Added comments and cleaned after review
